pyproject.toml
[project]
name = "ai-research-assistant-using-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "faiss-cpu>=1.13.2",
    "google-genai>=1.67.0",
    "google-generativeai>=0.8.6",
    "langchain>=1.2.12",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.19",
    "langchain-huggingface>=0.1.0",
    "langchain-text-splitters>=0.3.0",
    "pypdf>=6.9.1",
    "python-dotenv>=1.2.2",
    "sentence-transformers>=5.3.0",
    "streamlit>=1.55.0",
    "tiktoken>=0.12.0",
]
